[CCUBE-1541][JY] Combine config hooks. Edit element basic details test cases.
