+++ src/components/element-editor/basic-details/elements/email/email-basic-details.tsx
@@ -10,7 +10,7 @@ import {
     TypeField,
 } from "../../common";
 
-const fields = [
+const FIELDS = [
     { name: "type", component: TypeField },
     { name: "label", component: LabelField },
     { name: "required", component: MandatoryField },
@@ -26,7 +26,7 @@ const fields = [
 export const EmailBasicDetails = () => {
     return (
         <div data-testid="email-basic-details">
-            {fields.map(({ name, component: Component }) => {
+            {FIELDS.map(({ name, component: Component }) => {
                 const shouldShow = useShouldShowField(
                     name,
                     EElementLabel.EMAIL
