+++ src/context-providers/config/config-hook.ts
@@ -1,7 +1,22 @@
 import { useContext } from "react";
+import { EElementLabel } from "src/data";
 import { ConfigContext } from "./config-provider";
 
 export const useConfigContext = () => {
     const context = useContext(ConfigContext);
     return context;
+};
+
+export const useShouldShowField = (
+    fieldName: string,
+    elementName: EElementLabel
+) => {
+    const { elements, attributes } = useConfigContext();
+    const elementConfig = elements?.[elementName];
+
+    const shouldShow =
+        elementConfig?.attributes?.[fieldName]?.shouldShow ??
+        attributes?.[fieldName]?.shouldShow ??
+        true;
+    return shouldShow;
 };
