+++ src/context-providers/config/index.ts
@@ -1,4 +1,3 @@
 export * from "./config-hook";
 export * from "./config-provider";
-export * from "./hook";
 export * from "./types";
