+++ tests/components/element-editor/basic-details/elements/contact/contact-basic-details.spec.tsx
@@ -36,7 +36,7 @@ describe("ContactBasicDetails", () => {
 
         it.each(attributeFields)(
             "should hide the %s field when shouldShow in attributes is set to false",
-            async (fieldName) => {
+            (fieldName) => {
                 renderComponent({
                     configContext: {
                         attributes: {
